+++ DekorTrade99/src/hu/dekortrade99/server/jdo/Cikk.java
@@ -13,7 +13,7 @@ public class Cikk {
 	private String megnevezes;
 
 	@Persistent
-	private Float ar;
+	private Double ar;
 
 	@Persistent
 	private Integer kiskarton;
@@ -22,16 +22,16 @@ public class Cikk {
 	private Integer darab;
 
 	@Persistent
-	private Float terfogat;
+	private Double terfogat;
 
 	@Persistent
 	private String jel;
 
 	@Persistent
-	private Float bsuly;
+	private Double bsuly;
 
 	@Persistent
-	private Float nsuly;
+	private Double nsuly;
 
 	@Persistent
 	private Integer kepek;
@@ -39,9 +39,9 @@ public class Cikk {
 	@Persistent
 	private Boolean torolt;
 
-	public Cikk(String cikkszam, String megnevezes, Float ar,
-			Integer kiskarton, Integer darab, Float terfogat, String jel,
-			Float bsuly, Float nsuly, Integer kepek, Boolean torolt) {
+	public Cikk(String cikkszam, String megnevezes, Double ar,
+			Integer kiskarton, Integer darab, Double terfogat, String jel,
+			Double bsuly, Double nsuly, Integer kepek, Boolean torolt) {
 		this.cikkszam = cikkszam;
 		this.megnevezes = megnevezes;
 		this.ar = ar;
@@ -71,11 +71,11 @@ public class Cikk {
 		this.megnevezes = megnevezes;
 	}
 
-	public Float getAr() {
+	public Double getAr() {
 		return ar;
 	}
 
-	public void setAr(Float ar) {
+	public void setAr(Double ar) {
 		this.ar = ar;
 	}
 
@@ -95,11 +95,11 @@ public class Cikk {
 		this.darab = darab;
 	}
 
-	public Float getTerfogat() {
+	public Double getTerfogat() {
 		return terfogat;
 	}
 
-	public void setTerfogat(Float terfogat) {
+	public void setTerfogat(Double terfogat) {
 		this.terfogat = terfogat;
 	}
 
@@ -111,19 +111,19 @@ public class Cikk {
 		this.jel = jel;
 	}
 
-	public Float getBsuly() {
+	public Double getBsuly() {
 		return bsuly;
 	}
 
-	public void setBsuly(Float bsuly) {
+	public void setBsuly(Double bsuly) {
 		this.bsuly = bsuly;
 	}
 
-	public Float getNsuly() {
+	public Double getNsuly() {
 		return nsuly;
 	}
 
-	public void setNsuly(Float nsuly) {
+	public void setNsuly(Double nsuly) {
 		this.nsuly = nsuly;
 	}
